Update patch set 1
Patch Set 1:

Yes, you're correct. This only changes the default, and inverts the logic of the thing:

We were defaulting to the generic APN, and forcing people to manually choose the carrier's own default. We now default to the most common case for these terminals (remember: these are carrier-branded variants. This patch should rarely, if ever, be applied to open-market devices)

Patch-set: 1