Update patch set 1
Patch Set 1:

It will still choose the LTE apn as the default. I'm using a non-LTE SIM in my Note right now. This means I'll have to manually select the wrap.cingular apn after a clean install or resetting the apn list. This isn't normal though. The LTE apn is the common use case and that's what we want.

Patch-set: 1