Update patch set 1
Patch Set 1:

Steve, true this patch does fix the most used case.

An alternative: The application configuration on a LTE sim is different from a non-LTE sim.  So, we could detect the difference in the RIL and choose the proper APN accordingly.  However, this is pretty specific to AT&T.  I don't know how to make this algorithm general.  I assume this alternative approach is not acceptable?

Patch-set: 1