Update patch set 1
Patch Set 1:

The Qualcomm superclass already knows if you have an LTE SIM or not (the mUSIM boolean), and you could export that.

I'd advise against it, though. A user will usually know that his/her plan doesn't support LTE, so it isn't hard to explain having to reconfigure a device they bought with the knowledge of it having that capability (and probably discarded the bundled SIM card in the process).

And technically, even having an LTE-capable SIM does NOT imply having an LTE plan subscribed. I personally have 4 USIM, and only one of them has an LTE pla. Basing this decision on the technology of the card is actually wrong :)

Patch-set: 1