Update patch set 1
Patch Set 1:

Unfortunately, with this change and the TelephonyProvider overlay in device, Skyrocket still can't send MMS unless I delete the wap.cingular APN.

The reason seems to be that GsmDataConnectionTracker::setupData() does not consult TelephonyProvider when it picks an APN for mms.  So this patch is ineffective.  It just uses apnContext.getNextWaitingApn() which picks the first one in the list, wap.cingular.

Since the data connection is made properly, setupData() stops scanning and never makes it to the correct APN (pta).  But sending of the mms just hangs.

See this radio log snippet.  It should choose the pta APN (my selected APN in Settings), not wap.cingular. http://pastie.org/3953121

I have the pta APN selected in Settings.  If we could get GsmDataConnectionTracker::setupData() to honor this setting, I believe mms would work properly.

I do not know the code well enough to get a handle to TelephonyProvider and call getPreferredApnId().  Or perhaps there is a more proper fix.  Any suggestions you have would be helpful.

Patch-set: 1