Update patch set 1
Patch Set 1:

Hmm, I'm not sure what the best way to do the normalization is then. Don't do the keypad mapping, but accept anything accepted by GsmAlphabet.charToGsm()? If yes, does that work with CDMA SMS? And even if it does, the point of the normalization is to strip out any formatting characters (e.g. '(', ')', ' ', '-'), which all are perfectly encodable in GSM-7bit.

I'm currently not sure whether this really is a problem in practise. Non-alphanumeric characters are simply dropped in the normalization, so a name like 'ÆBCDEFG' should work fine; it's just that a name like '?ì?ò?' won't work due to all characters being rejected in the normalization. Is this a real issue?

Patch-set: 1
Reviewer: Danny Baumann <1063@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0