Update patch set 4
Patch Set 4:

Yes, but they still can't change public API, even if they might have valid reasons. It's - AFAICT - impossible to map sub ID to phone ID for all cases (only mapping it in the projection like this patch does is wrong, one must also handle the selection), so we shouldn't attempt that and - if their commit message is right, which I'm not sure about because I don't think we filter by sub ID - deal with it in the Mms app.

Patch-set: 4