Update patch set 1
Patch Set 1:

Yes, the data migration is crucial; we don't require our users to wipe data on new releases.
Getting the phone ID for a given sub ID is simple (simple lookup in the SIM info table); the other way not so much, as it requires the SIMs to come up first, which happens later than the telephony provider starting up.

And yes, the sub ID is important, otherwise it wouldn't be part of the API contract :) Whether to store phone ID (basically, the SIM slot #) or the subscription ID (an identifier of the SIM card) is probably a matter of perspective, but as sub ID is part of the API contract, arguing which of both is more correct is pointless and we _have_ to do the conversion. CAF didn't do it either because they didn't notice the API breakage, or they didn't care, or just aren't ready yet.

Patch-set: 1