Update patch set 3
Patch Set 3:

The new messaging app in M has a local database for blacklist.  We also have the concept of a blacklist from old mms app which is stored in telephony provider.  The idea was to synchronize the data (which the new M messaging app already does) and we piggy backed onto that synchronization task and synchronize the black list.  This should definitely be a good design for the upgrade path.  what would you propose instead?

Patch-set: 3