Update patch set 3
Patch Set 3:

I should probably expand a bit on my last statement.

The thing I am opposed is having a _persistent_ cache of the telephony blacklist in the Messaging app - that's what generates the complicated sync infrastructure. If we do need the blacklist data in the Messaging content provider, AFAICT there are other ways to achieve that, e.g. having an in-memory cache of the data, updated whenever there's a change notification for the blacklist data. I realize this is the more involving approach, especially if the blacklist column appears in where clauses, but it should be possible nevertheless and is significantly cleaner from a DB structure perspective.

My -2 also only is for the timestamp column, the notification URI stuff is fine.

Patch-set: 3