Update patch set 3
Patch Set 3:

Thanks for the explanation.  I think we are deciding to go with synchronizing the archived bit based on whether it was blocked in the provider.  However, since our observer will not be running 100% of the time (i.e. process is dead) we would need the time stamp in order to know whether we need to handle the sync process on startup again.  After that then the timestamp is useless.  Do you have any input or ideas on a way to handle that use case?

Patch-set: 3