Update patch set 3
Patch Set 3:

I experimented with how long it takes to synchronize fw db to local db (same operation we would have to do at startup if we want to update the archived bit, instead of the blocked bit). For 50 blocked numbers in the fw db, it takes about 950ms to update the blocked bit in the local db participants table. So, if there are 50 conversations, it will probably take about that long to update the archived bit. This happens in the background. Seems reasonable to me. If we go with the approach of always updating the archived bit at startup, based on the blocked bit in the fw db, then we wouldn't need the timestamp functionality.

Patch-set: 3