Update patch set 3
Patch Set 3:

After working on the proposed implementation, I've realized that the participant's blocked status is checked in several places and each requires a query into the FW blacklist DB and then a query into the Participants Table, which results in messier code and a bigger performance hit for frequent operations. I'd like to propose a slightly different implementation.

1. eliminate FW blacklist DB timestamp
2. treat the Participant.Blocked bit as a write through cached version of the FW blacklist DB
3. at startup time, always synchronize the Participant.Blocked bit to the FW Blacklist DB. The previous plan was to update the Conversation.Archived bit, instead will update the Participant.Blocked bit. The performance impact is negligible based on some experimentation.
4. the rest is as in the original proposal, where block/unblock actions update the Participant.Blocked and since it's treated as a write through type cache, the FW Blacklist DB will be also updated.
5. a ContentObserver will monitor the FW Blacklist DB to keep the Participant.Blocked updated when changes are made from the FW.

As an alternative to treating the Participant.Blocked as a write through cache, an in memory write through cache can be created for the FW Blacklist DB. This approach will result in more code refactoring and the messaging code will not be as clean, as it is now optimized for using Participant.Blocked when making queries into the Participants table for blocked/unblocked determinations. From a performance point of view, the impact is insignificant, since the Participants table has to be queried regardless of how the Blocked state is implemented (in memory or in the Participants table itself). And the updates to the Participant.Blocked are very rare, so it doesn't matter if the update goes to the memory cache or the DB.

Patch-set: 3