Update patch set 2
Patch Set 2:

Matt,
I made the requested fixes. Sorry about the file mode change - good catch. I edited it over a network share and it must have changed the mode. I'll double check settings and be cautious of that, and pay more attention to code style :-)

As far as where the regression occurred, it was when M was merged in 0986a3939dae624163bf7df5bbe298389a1d16e2. 

The original commit that I based restoring this off of was for L in c51b725f3d3b598857aab6204f97a2f049c7017a

Thanks for the review and let me know if anything else is needed.

Also: from further testing, it's worth noting that this current sets the APN after the modem is started, so it doesn't take effect immediately. To take effect I either have to toggle airplane or reboot. 

This seems to be an general issue with APN changing right now, at least on my device (always needs a reboot or an airplane toggle if I toggle them from the APN menu, but nothing tells me that). 

I'm not sure how to best fix this, and I am open to suggestion as to what is appropriate, but it's likely outside the scope of this change. Here, we at least get the user bound to the right initial APN if this setting is invoked so that they are set for future use.

Lot of devices are having LTE association issues right now, including my l900. When I brought this up on IRC I was linked to here: http://review.cyanogenmod.org/#/c/150149/

Using this type of setting at least feels like a better way than nuking MVNOs out of the apn settings, but if there is a better long term fix I'm also open to suggestion. In the meantime, this seems to be a pretty safe way to indicate a specific apn for devices that are sold from a specific carrier.

Patch-set: 2