Update patch set 2
Patch Set 2:

It must be doing something because the changes in the db-upgrade topic fix all the FCs I otherwise get. I suppose it could simply be that it successfully updates the db version and that being enough to make the apps happy because the functionality using the extra database fields has not yet been implemented.

Patch-set: 2